requirements.txt
blinker==1.4
cycler==0.11.0
dill==0.3.5
fonttools==4.33.3
graphviz==0.20
kiwisolver==1.4.2
matplotlib==3.5.2
networkx==2.8.1
nose==1.3.7
numpy==1.22.3
oemof-visio @ git+https://github.com/oemof/oemof_visio.git@8830fd34db3b4c0695f63d617e9346294e8b1c47
oemof.network==0.4.0
oemof.solph==0.4.4
oemof.tools==0.4.1
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
ply==3.11
pydantic==1.9.1
Pyomo==5.7.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyUtilib==6.0.0
six==1.16.0
typing_extensions==4.2.0
